Tutorial v2
+ Tutorial script is now saving scored points, player state and player tilt
+ New sound effects added
+ Placeholder tutorial screens added
+ Small localization bug fixed

+++ PolskaGraNarodowaRemake/Assets/Scripts/PlaneScript.cs
@@ -53,7 +53,7 @@ internal class Plane
     internal bool attackKeyReleased;
     internal bool activeBottleWarning;
     //Score
-    internal int gameScore;
+    internal float gameScore;
     internal bool rewardForLandingAdded;
     //PowerUps
     internal bool shieldEnabled;
